requirements.txt
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.7.0
PyYAML>=6.0
tqdm>=4.65.0
transforms3d>=0.4.1
matplotlib>=3.5.0
opencv-python>=4.7.0